+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project are documented in this file.
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased 3.x](https://github.com/opensearch-project/anomaly-detection/compare/3.4...HEAD)
+
+
 ### Features
+- Correlating Anomalies via Temporal Overlap Similarity ([#1641](https://github.com/opensearch-project/anomaly-detection/pull/1641))
+
 ### Enhancements
 ### Bug Fixes
 ### Infrastructure